meta-ostro: upgrade linux-firmware to enable 7260 wifi chip for quark
Linux kernel 4.0 is used in the quark image of ostro.
The old linux-firmware doesn't match with linux kernel 4.0.
Some firmwares of linux kernel 4.0 are missed in old linux-firmware.

It results in that Intel 7260 wifi chip can't be enabled.
Intel 7260 wifi chip can't work well in the quark image of ostro.

Therefore, upgrade linux-firmware to enable 7260 wifi chip for quark.

Signed-off-by: Wu Zheng <wu.zheng@intel.com>