edison.conf: prepare for new way of increasing Ostro image sizes
This overrides a variable that ostro-image will use to increase the image
size only if the machine definition does not set something itself.

A better solution would be to control the amount of space added by
rootfs_rpm.bbclass, but that depends on changing OE-core, which is not
an option at this point in time.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>