package selection: via machine config
A BSP should not have to make assumptions about the distro and/or
image it is used with. The mechanism for adding machine specific
packages to images are MACHINE_ESSENTIAL_EXTRA_RRECOMMENDS and
MACHINE_ESSENTIAL_EXTRA_RDEPENDS.

The bcm43340 modules where already listed. It is unclear why
it was necessary to also add them explicitly to IMAGE_INSTALL;
the mechanism seems to work fine now.

bluetooth-rfkill-event was missing. It's now declared as hard
dependency, i.e. the image will not build without it.

With that in place, the ostro-image.bbappend is not longer necessary
and can be removed, which is important because Ostro is going to
introduce other image recipes.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>