ostro-image.bbappend: replace by MACHINE_ESSENTIAL_EXTRA_RDEPENDS in layer.conf
We are about to have multiple different Ostro images, so a single
.bbappend would not be good enough anymore. Better use the normal
mechanism, extending MACHINE_ESSENTIAL_EXTRA_RDEPENDS, for installing
additional packages.

We have to do that in the layer.conf with machine overrides because we cannot
extend the upstream machine definitions. We could do it for Edison, but
for the sake of consistency, systemd-watchdog is also added to the layer.conf.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>