Modify 802.15.4 module to match load TI and Atmel chips drivers
Modify 802.15.4 module to match load multi chips.
The chips include TI and Ateml 802.15.4 chips.
Atmel 802.15.4 chips are added and supported.
At current, Atmel 802.15.4 chips are used for MinnowBoard Max.
The Atmel 802.15.4 chips need to be supported in Minnowboard Max.
The kernel modules are implemented to enable IRQ of the chip,
and enable spi interface of the chip.

Signed-off-by: Wu zheng <wu.zheng@intel.com>