linux-yocto: add a kernel config fragment required by bootchart
This commit adds a kernel config fragment that enables kernel
configuration options required by systemd-bootchart. This config
fragment is *not* used by default. It can be enabled by adding it
via a local.conf line such as:
   SRC_URI_append_pn-linux-yocto = " file://bootchart.cfg"
Or can also be added to the recipe file (.bb or .bbapend file)

Signed-off-by: Geoffroy Van Cutsem <geoffroy.vancutsem@intel.com>