u-boot-osip.inc: Python3 compatibility
Strings no longer have the "decode" method in Python3, and support for
the "hex" encoding is a bit spotty for binary strings. The binascii
approach works the same for Python2 and Python3, so that gets used
here.

Python3 also changes the semantic of the / operator: the result is
always float. To avoid the floating point division, the desired result
is spelled out explicitly.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>