linux-yocto_4.4.bbappend: Add usb-ethernet drivers to all images
Some, but not all usb-ethernet adapters work with images that target other
devices than Edison. Adding the drivers that currently are used only on
Edison images expands the set of functional adapters. This can be
useful, as some non-Edison boards require usb-ethernet adapters as
well. Currently finding a functional adapter is rather painful.

Signed-off-by: Erkka Kääriä <erkka.kaaria@intel.com>