u-boot: drop the SD/eMMC boot partition patch
The latest oe-core brings u-boot 2016.11.bb so bitbake fails the
missing matching recipe for 2016.03.bbappend.

The BBB patch no longer applies on top of 2016.11 but it also
seems the patch is not needed anymore because of
"ti_armv7_common: env: Fix hard coded mmc device for uuid" fix
in it.

Drop the modification to select between eMMC/SD.

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>